Add support for --baseline to browser-specific-failures.js (#22)
This command line flag causes the results to be filtered to only include
tests that existed on the baseline date. In essence this is a rough form
of time-travel, allowing us to backdate a modern WPT run to a previous
date, to better see how a browser has progressed over time (by ignoring
newly added tests).