[compat2021] Add a tooltip action for version diffing
This adds a tooltip action to open a revision diff view for the given
tooltip. Currently this works only for Google Chrome, but is shown for
all three browsers as the charts library we are using doesn't have great
support for figuring out what tooltip is visible to the user*.

See https://github.com/Ecosystem-Infra/wpt-results-analysis/issues/34

* Even worse, the action is global but technically you can have two
tooltips showing (one selected and one hover), and you might want to
show the action for one tooltip but not the other. Completely impossible
with Google Charts!