[compat2021] Add tooltips for the summary numbers
This gives a quick overview of which areas are the good or concerning
ones for each product. The experience isn't super fluid, but it works
for now and we can build on it.

As part of this, I noticed that we were accidentally rounding pass rates
of 90%+ up to 100% when calculcating the summary scores, which biases
for browsers that are close to full pass rate. Switch to floor instead.