Remove houdini-dependent test from compat2021 set (#66)
The WPT test transform-interpolation-computed-value.html is largely about testing a CSS Houdini API, rather than about testing CSS Transforms.  The top of the test references Houdini as the spec being tested:
```
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om/#transformvalue-objects">
```

At this point, the CSS Houdini APIs are only implemented in Chromium. So this test only passes in Chromium, due to its dependence on that.

It's not clear that Houdini itself is a source of interop issues at this point, so it doesn't make a lot of sense to include houdini-dependent tests in the compat2021 test set. Therefore, I propose we remove this test from the test set.