Simplify scoreSubtests
We don't need the early-exit of the loop, as long as the other side is in an
else block, which means we then have a single place where we advance subtests,
making it easier to reason about the behaviour.

With that done, we can easily move to having a single check for prevName.