Fix #126: be slightly forgiving on dates of runs
Previously, aligned runs were only found if they have a time_start
value that falls within the date being queried for, and if one or more
of the runs that would be considered aligned started outside of this,
then those runs will not be considered aligned and returned.

Instead, allow runs which started either on the date being queried for
or the day before, and treat the returned runs as a return for the
appropriate date provided at least one run started on that date
(thereby avoiding returning a run which was valid for the previous
day).

We could increase the amount of fuzziness allowed here, but it
increases the probability of ending up with a significantly mismatched
run appearing out-of-order.