Update web-platform-tests/results-analysis-cache with all products
We might want to start scoring mobile browsers (such as Android WebView,
Chrome for Android, Chrome for iOS, or Firefox for Android) in the
future, and back-filling the cache is painfully slow (as it requires
downloading each run). Instead, let's just add every product to the
cache, because it doesn't have a particularly significant affect on the
overall size of the repo.