Update browser scoring for Servo
While trying to compute the score of Servo, I ran into the issue that
tests reported as "OK" would crash. This is a valid test status as
reported by the API [1].

Additionally, while trying to plot the scores, I saw that all numbers
are shown as floats. This makes analysis a bit more difficult and
appears to also show on the dashboard on https://wpt.fyi/results/

Therefore, floor all scores to result in whole numbers.

[1]: https://github.com/web-platform-tests/wpt.fyi/blob/main/api/README.md#apiresults